+++ redis-monitor/redis_monitor.py
@@ -3,6 +3,7 @@ import sys
 import time
 import traceback
 import argparse
+import time
 
 from collections import OrderedDict
 from scutils.log_factory import LogFactory
@@ -215,7 +216,7 @@ class RedisMonitor:
 
     def _setup_stats_plugins(self):
         '''
-        Sets up the total stats collectors
+        Sets up the plugin stats collectors
         '''
         self.stats_dict['plugins'] = {}
         for key in self.plugins_dict:
@@ -251,6 +252,7 @@ class RedisMonitor:
 
         @param item: the unique print for HLL counter
         '''
+        item = item + str(time.time())
         if 'total' in self.stats_dict:
             self.logger.debug("Incremented total stats")
             for key in self.stats_dict['total']:
@@ -265,6 +267,7 @@ class RedisMonitor:
 
         @param item: the unique print for HLL counter
         '''
+        item = item + str(time.time())
         if 'fail' in self.stats_dict:
             self.logger.debug("Incremented fail stats")
             for key in self.stats_dict['fail']:
@@ -280,6 +283,7 @@ class RedisMonitor:
         @param name: The formal name of the plugin
         @param item: the unique print for HLL counter
         '''
+        item = item + str(time.time())
         if 'plugins' in self.stats_dict:
             self.logger.debug("Incremented plugin '{p}' plugin stats"\
                     .format(p=name))